refactor(pi): extract shared RPC transport core into pi-shared
Move the Pi CLI runtime, RPC types, session descriptor, tool-call mapper,
history mapper, rewind, and fake-pi test utility into providers/pi-shared,
parameterized by a dialect contract. The pi provider becomes a thin facade
over the shared core; behavior is gated by the existing Pi e2e suite.

Intentional Pi-visible deltas (enumerated):
- slash commands Pi handles without starting a turn now surface command
  output and complete the turn locally
- persisted custom messages replay through the history mapper
