+++ packages/server/src/server/agent/providers/pi-shared/runtime.ts
@@ -13,25 +13,31 @@ export interface PiRuntimeLaunch {
   cwd: string;
   argv: string[];
   env?: Record<string, string>;
+  protocolMode?: "rpc" | "rpc-ui";
   model?: string;
   thinkingOptionId?: string;
+  modeId?: string;
   session?: string;
   noSession?: boolean;
   systemPrompt?: string;
   mcpConfigPath?: string;
   extensionPaths?: string[];
+  extraArgs?: string[];
 }
 
 export interface PiStartSessionInput {
   cwd: string;
   env?: Record<string, string>;
+  protocolMode?: "rpc" | "rpc-ui";
   model?: string;
   thinkingOptionId?: string;
+  modeId?: string;
   session?: string;
   noSession?: boolean;
   systemPrompt?: string;
   mcpConfigPath?: string;
   extensionPaths?: string[];
+  extraArgs?: string[];
 }
 
 export interface PiRuntimeSession {
@@ -50,6 +56,8 @@ export interface PiRuntimeSession {
   setThinkingLevel(level: string): Promise<void>;
   getSessionStats(): Promise<PiSessionStats>;
   getCommands(): Promise<PiRpcSlashCommand[]>;
+  request(command: { type: string; [key: string]: unknown }, timeoutMs?: number): Promise<unknown>;
+  sendRawFrame(frame: object & { type: string }): void;
   respondToExtensionUiRequest(
     id: string,
     response: { value?: string; confirmed?: boolean; cancelled?: boolean },
@@ -73,30 +81,9 @@ export function buildPiLaunch(input: {
       : input.command;
   const argv = [...command];
 
-  if (!hasModeRpc(argv)) {
-    argv.push("--mode", "rpc");
-  }
-  if (input.session.model) {
-    argv.push("--model", input.session.model);
-  }
-  if (input.session.thinkingOptionId) {
-    argv.push("--thinking", input.session.thinkingOptionId);
-  }
-  if (input.session.noSession) {
-    argv.push("--no-session");
-  } else if (input.session.session) {
-    argv.push("--session", input.session.session);
-  }
+  const protocolMode = input.session.protocolMode ?? "rpc";
   const systemPrompt = input.session.systemPrompt?.trim();
-  if (systemPrompt) {
-    argv.push("--append-system-prompt", systemPrompt);
-  }
-  if (input.session.mcpConfigPath) {
-    argv.push("--mcp-config", input.session.mcpConfigPath);
-  }
-  for (const extensionPath of input.session.extensionPaths ?? []) {
-    argv.push("--extension", extensionPath);
-  }
+  appendPiLaunchArgs(argv, input.session, protocolMode, systemPrompt);
 
   return {
     cwd: input.session.cwd,
@@ -110,20 +97,57 @@ export function buildPiLaunch(input: {
         : undefined,
     model: input.session.model,
     thinkingOptionId: input.session.thinkingOptionId,
+    protocolMode,
+    modeId: input.session.modeId,
     session: input.session.session,
     noSession: input.session.noSession,
     systemPrompt,
     mcpConfigPath: input.session.mcpConfigPath,
     extensionPaths: input.session.extensionPaths,
+    extraArgs: input.session.extraArgs,
   };
 }
 
-function hasModeRpc(argv: string[]): boolean {
+function appendPiLaunchArgs(
+  argv: string[],
+  session: PiStartSessionInput,
+  protocolMode: "rpc" | "rpc-ui",
+  systemPrompt: string | undefined,
+): void {
+  if (!hasModeFlag(argv)) {
+    argv.push("--mode", protocolMode);
+  }
+  if (session.extraArgs?.length) {
+    argv.push(...session.extraArgs);
+  }
+  if (session.model) {
+    argv.push("--model", session.model);
+  }
+  if (session.thinkingOptionId) {
+    argv.push("--thinking", session.thinkingOptionId);
+  }
+  if (session.noSession) {
+    argv.push("--no-session");
+  } else if (session.session) {
+    argv.push("--session", session.session);
+  }
+  if (systemPrompt) {
+    argv.push("--append-system-prompt", systemPrompt);
+  }
+  if (session.mcpConfigPath) {
+    argv.push("--mcp-config", session.mcpConfigPath);
+  }
+  for (const extensionPath of session.extensionPaths ?? []) {
+    argv.push("--extension", extensionPath);
+  }
+}
+
+function hasModeFlag(argv: string[]): boolean {
   for (let i = 0; i < argv.length; i += 1) {
-    if (argv[i] === "--mode" && argv[i + 1] === "rpc") {
+    if (argv[i] === "--mode") {
       return true;
     }
-    if (argv[i] === "--mode=rpc") {
+    if (argv[i]?.startsWith("--mode=")) {
       return true;
     }
   }
