+++ packages/server/src/server/agent/providers/pi/rewind.ts
@@ -1,14 +1,2 @@
-export interface PiRewindNavigator {
-  navigateTree(targetId: string): Promise<unknown>;
-}
-
-export async function revertPiConversation(input: {
-  messageId: string;
-  navigator: PiRewindNavigator;
-}): Promise<void> {
-  const targetId = input.messageId.trim();
-  if (!targetId) {
-    throw new Error("Pi rewind requires a user message id");
-  }
-  await input.navigator.navigateTree(targetId);
-}
+export { revertPiConversation } from "../pi-shared/rewind.js";
+export type { PiRewindNavigator } from "../pi-shared/rewind.js";
