+++ packages/server/src/server/agent/providers/pi/tool-call-mapper.test.ts
@@ -5,7 +5,7 @@ import {
   parseToolArgs,
   parseToolResult,
   resolveToolCallName,
-} from "./tool-call-mapper.js";
+} from "../pi-shared/tool-call-mapper.js";
 
 describe("Pi tool call mapper", () => {
   test("maps bash args and result to shell detail", () => {
